BUGFIX: Use different destination directory based on the source number
If you have muliple sources, that go into the same destination
directory or specify a source mulitple times on the command line
and the backup is done within a minute, the destination directory
will still be the name:

   497    export destination_name="${INTERVAL}.$(${CDATE}).$$"

None of the values will change for the next backup. Thus the new
format since ccollect 0.8.1 is

   497    export destination_name="${INTERVAL}.$(${CDATE}).$$-${source_no}"

which contains the number of the current source.

Signed-off-by: Nico Schottelius <nico@ikn.schottelius.org>